Environment Shits
Fixed the Lighting on the lantern to cover a little more area, and it doesn't blind the player. Adjusted the size of the evidence tapes to be more of a realistic VHS tape. Modified the textures of the lantern to allow light to pass through—added props around the 1st floor to direct the player.