Animation Offset
Added an animation offset primarily for the flashlight idle animation when the character looks up or down, but it is active for all other animations as well. Renamed some assets in the animation folder, e.g - AS_Idle (AS - Animation Sequence). Parented the first person camera to the character mesh and attached it to the head via parent socket. (The camera now moves with the character's head instead of floating stationary. I think this works for SITD to further enhance realism and immersion)